test_bft_DNV/requirements.txt
pytest==
selenium==
allure-pytest==
webdriver-manager==